use common fn for inputs
